download large files working(previous- wasnt able to allocate large memory on stack, so changed to heap)

+++ torrentclient/fileHandler.h
@@ -16,7 +16,7 @@ class FileHandler
     std::string getFileHash(std::string file_name);
     long fileSize(std::string path);
     void readFileChunk(int chunk_index, std::string path, std::vector<char>& buf_b);
-    void createEmptyFile(std::string path, int size);
+    void createEmptyFile(std::string path, long size);
 };
 
 #endif
